+++ src/ray/python/buffer_writer.h
@@ -28,8 +28,8 @@ int RawDataWrite(const std::string &raw_data,
 class PythonObjectBuilder {
  public:
   PythonObjectBuilder();
-  void AppendBuffer(uint8_t *buf, int64_t length, int64_t itemsize, int32_t ndim,
-                    char* format, int64_t *shape, int64_t *strides);
+  void AppendBuffer(uint8_t *buf, ssize_t length, ssize_t itemsize, int ndim,
+                    char* format, ssize_t *shape, ssize_t *strides);
   // Get the total size of the current builder.
   int64_t GetTotalBytes();
   int64_t GetInbandDataSize() { return inband_data_size; };
